+++ docs/web_services/examples/shell/SuggestedName/showSuggestedName.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+
+url=$1
+username=$2
+password=$3
+suggested_name=$4
+
+usage() {
+    echo "Sample script for showing statuses via web services"
+    echo "Usage:    ./showSuggestedName.sh <complete_apollo_URL> <username> <password> <suggested_name>"
+    echo "Example:  ./showSuggestedName.sh http://localhost:8080/apollo demo demopass resolved"
+    echo ""
+}
+
+if [[ ! -n "$url" || ! -n "$username" || ! -n "$password"  || ! -n "$suggested_name" ]]; then
+    usage
+    exit
+fi
+
+echo curl -H 'Content-type: application/json' -X POST ${url}/suggestedName/showName -d "{'username': '${username}', 'password': '${password}',name:'${suggested_name}'}"
+curl -H 'Content-type: application/json' -X POST ${url}/suggestedName/showName -d "{'username': '${username}', 'password': '${password}',name:'${suggested_name}'}"
+
+