+++ grails-app/conf/org/bbop/apollo/SecurityFilters.groovy
@@ -2,12 +2,16 @@ package org.bbop.apollo
 
 import grails.converters.JSON
 import org.apache.shiro.SecurityUtils
-import org.apache.shiro.authc.UsernamePasswordToken
 import org.apache.shiro.subject.Subject
+import org.codehaus.groovy.grails.web.json.JSONObject
+import org.springframework.http.HttpStatus
 
 class SecurityFilters {
 
     def permissionService
+
+    // NOTE: this is used for multiple security filters
+   public static final List<String> WEB_ACTION_LIST = ['index','show','create','edit','update','delete']
 
     def filters = {
 
@@ -79,7 +83,15 @@ class SecurityFilters {
 //                                    paramString = paramString.substring(indexOfLoc)
 //                                }
                                 targetUri = targetUri + paramString
-                                if (paramString.contains("http://") || paramString.contains("https://") || paramString.contains("ftp://")) {
+
+                              // respond to JSON this way
+                              if(request.JSON?.size()>0){
+                                response.status = HttpStatus.UNAUTHORIZED.value()
+                                render new JSONObject("error":"Failed to authenticate")
+                                return false
+                              }
+                              else
+                              if (paramString.contains("http://") || paramString.contains("https://") || paramString.contains("ftp://")) {
                                     redirect(uri: "${request.contextPath}/auth/login?targetUri=${targetUri}")
                                 }
                                 else {