+++ grails-app/controllers/org/bbop/apollo/ProvenanceController.groovy
@@ -2,6 +2,7 @@ package org.bbop.apollo
 
 import grails.converters.JSON
 import grails.transaction.Transactional
+import org.bbop.apollo.gwt.shared.GlobalPermissionEnum
 import org.bbop.apollo.gwt.shared.PermissionEnum
 import org.bbop.apollo.history.FeatureOperation
 import org.codehaus.groovy.grails.web.json.JSONArray
@@ -14,6 +15,7 @@ import org.restapidoc.pojo.RestApiParamType
 import org.restapidoc.pojo.RestApiVerb
 
 import static org.springframework.http.HttpStatus.NOT_FOUND
+import static org.springframework.http.HttpStatus.UNAUTHORIZED
 
 @RestApi(name = "Provenance Annotation", description = "Methods for managing provenance annotations")
 @Transactional(readOnly = true)
@@ -34,7 +36,10 @@ class ProvenanceController {
   )
   def index() {
     JSONObject dataObject = permissionService.handleInput(request, params)
-    permissionService.checkPermissions(dataObject, PermissionEnum.READ)
+    if(!permissionService.checkLoginGlobalAndLocalPermissions(dataObject, GlobalPermissionEnum.USER,PermissionEnum.READ)){
+      render status : UNAUTHORIZED
+      return
+    }
     Feature feature = Feature.findByUniqueName(dataObject.uniqueName as String)
     if (feature) {
       JSONObject annotations = provenanceService.getAnnotations(feature)
@@ -68,7 +73,10 @@ class ProvenanceController {
   @Transactional
   def save() {
     JSONObject dataObject = permissionService.handleInput(request, params)
-    permissionService.checkPermissions(dataObject, PermissionEnum.WRITE)
+    if(!permissionService.checkLoginGlobalAndLocalPermissions(dataObject, GlobalPermissionEnum.USER,PermissionEnum.WRITE)){
+      render status : UNAUTHORIZED
+      return
+    }
     User user = permissionService.getCurrentUser(dataObject)
     Provenance provenance = new Provenance()
     Feature feature = Feature.findByUniqueName(dataObject.feature)
@@ -123,7 +131,10 @@ class ProvenanceController {
   @Transactional
   def update() {
     JSONObject dataObject = permissionService.handleInput(request, params)
-    permissionService.checkPermissions(dataObject, PermissionEnum.WRITE)
+    if(!permissionService.checkLoginGlobalAndLocalPermissions(dataObject, GlobalPermissionEnum.USER,PermissionEnum.WRITE)){
+      render status : UNAUTHORIZED
+      return
+    }
     User user = permissionService.getCurrentUser(dataObject)
     Feature feature = Feature.findByUniqueName(dataObject.feature)
 
@@ -172,7 +183,10 @@ class ProvenanceController {
   @Transactional
   def delete() {
     JSONObject dataObject = permissionService.handleInput(request, params)
-    permissionService.checkPermissions(dataObject, PermissionEnum.WRITE)
+    if(!permissionService.checkLoginGlobalAndLocalPermissions(dataObject, GlobalPermissionEnum.USER,PermissionEnum.WRITE)){
+      render status : UNAUTHORIZED
+      return
+    }
     User user = permissionService.getCurrentUser(dataObject)
 
     Feature feature = Feature.findByUniqueName(dataObject.feature)